Changes to R-factor and projection settings
R factor no longer uses point data and kriging but uses any annual rainfall raster that gives its values in mm. One such raster is the hadgrid 1km rainfall dataset available from the met office. This was done as the data is readily available from the met office and I assume this is true for other countries.

Additionally, because this tool is intended to be used in other countries, the CRS needs to be adaptable. So a code box was added where the user can choose the CRS they want the data to be displayed it and that is then applied across the workflow.